requirements.txt
click>=7.0
tox>=3.5
unidecode>=1.1.1
PyYAML>=5.3.1
dicttoxml>=1.7.4
pandas>=0.24.2
nltk>=3.4.1
spacy>=2.1.3
redis>=3.2.1
redisearch>=1.0.0
hiredis>=1.0.0
elasticsearch>=7.0.0,<8.0.0
pymongo>=3.10
pyarrow>=0.17.1
python-Levenshtein>=0.12
